move subscriptionOptions definition
